# src/docs/swagger.yaml
basePath: /v1/api
definitions:
  data.AdsConditions:
    properties:
      ageEnd:
        maximum: 100
        minimum: 1
        type: integer
      ageStart:
        maximum: 100
        minimum: 1
        type: integer
      country:
        items:
          $ref: '#/definitions/enum.CountryCode'
        type: array
      gender:
        $ref: '#/definitions/enum.Gender'
      platform:
        items:
          $ref: '#/definitions/enum.Platform'
        type: array
    required:
      - country
      - gender
      - platform
    type: object
  data.CreateAdsReq:
    properties:
      conditions:
        $ref: '#/definitions/data.AdsConditions'
      endAt:
        example: "2021-01-01T00:00:00Z"
        type: string
      startAt:
        example: "2021-01-01T00:00:00Z"
        type: string
      title:
        type: string
    required:
      - conditions
      - endAt
      - startAt
      - title
    type: object
  data.GetAdsResp:
    properties:
      EndAt:
        type: string
      Title:
        type: string
    type: object
  enum.CountryCode:
    enum:
      - CN
      - US
      - JP
      - TW
    type: string
    x-enum-varnames:
      - CN
      - US
      - JP
      - TW
  enum.Gender:
    enum:
      - M
      - F
      - B
    type: string
    x-enum-varnames:
      - Male
      - Female
      - Both
  enum.Platform:
    enum:
      - android
      - ios
      - web
    type: string
    x-enum-varnames:
      - Android
      - IOS
      - Web
info:
  contact: {}
  description: a simple API for advertisement
  title: Advertisement API
  version: "1.0"
paths:
  /ad:
    get:
      description: Gets a list of advertisements that match the specified conditions.
      parameters:
      - in: query
        maximum: 100
        minimum: 1
        name: age
        type: integer
      - enum:
          - CN
          - US
          - JP
          - TW
        in: query
        name: country
        required: true
        type: string
        x-enum-varnames:
          - CN
          - US
          - JP
          - TW
      - enum:
        - M
        - F
        - B
        in: query
        name: gender
        required: true
        type: string
        x-enum-varnames:
          - Male
          - Female
          - Both
      - in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      - in: query
        minimum: 0
        name: offset
        type: integer
      - enum:
        - android
        - ios
        - web
        in: query
        name: platform
        required: true
        type: string
        x-enum-varnames:
          - Android
          - IOS
          - Web
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/data.GetAdsResp'
            type: array
        "400":
          description: Bad Request
      tags:
      - Advertisement
    post:
      description: Creates a new advertisement with the specified title, start and
        end dates, and conditions.
      parameters:
      - description: Advertisement information
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/data.CreateAdsReq'
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
      tags:
      - Advertisement
swagger: "2.0"